server/requirements.txt
Flask
python-dotenv
langchain
langchain-core
